Refactoring for Main controller
