+++ src/main/java/com/tsel/home/project/booklibrary/controller/AbstractViewController.java
@@ -1,15 +1,11 @@
 package com.tsel.home.project.booklibrary.controller;
 
 import com.tsel.home.project.booklibrary.converter.BookConverter;
-import com.tsel.home.project.booklibrary.dao.repository.impl.AuthorRepository;
 import com.tsel.home.project.booklibrary.dao.repository.impl.AuthorRepositoryV2;
 import com.tsel.home.project.booklibrary.dao.repository.impl.BookRepository;
 import com.tsel.home.project.booklibrary.dao.repository.impl.BookRepositoryV2;
-import com.tsel.home.project.booklibrary.dao.repository.impl.CycleRepository;
 import com.tsel.home.project.booklibrary.dao.repository.impl.CycleRepositoryV2;
-import com.tsel.home.project.booklibrary.dao.repository.impl.PublisherRepository;
 import com.tsel.home.project.booklibrary.dao.repository.impl.PublisherRepositoryV2;
-import com.tsel.home.project.booklibrary.dao.repository.impl.UserSettingsRepository;
 import com.tsel.home.project.booklibrary.dao.repository.impl.UserSettingsRepositoryV2;
 import com.tsel.home.project.booklibrary.utils.elements.ButtonAnswer;
 import com.tsel.home.project.booklibrary.utils.elements.ImageProvider;
@@ -27,18 +23,10 @@ public abstract class AbstractViewController {
 
     public static final String RESOURCE_PATH = "/com/tsel/home/project/booklibrary/";
 
-    protected static final Logger LOGGER = LogManager.getLogger(AbstractViewController.class);
+    private static final Logger log = LogManager.getLogger(AbstractViewController.class);
 
     @Deprecated
     protected static final BookRepository BOOK_REPOSITORY = BookRepository.getInstance();
-    @Deprecated
-    protected static final CycleRepository CYCLE_REPOSITORY = CycleRepository.getInstance();
-    @Deprecated
-    protected static final AuthorRepository AUTHOR_REPOSITORY = AuthorRepository.getInstance();
-    @Deprecated
-    protected static final PublisherRepository PUBLISHER_REPOSITORY = PublisherRepository.getInstance();
-    @Deprecated
-    protected static final UserSettingsRepository USER_SETTINGS_REPOSITORY = UserSettingsRepository.getInstance();
 
     protected static final BookRepositoryV2 BOOK_REPOSITORY_V2 = BookRepositoryV2.INSTANCE;
     protected static final CycleRepositoryV2 CYCLE_REPOSITORY_V2 = CycleRepositoryV2.INSTANCE;
@@ -109,7 +97,7 @@ public abstract class AbstractViewController {
             stage.showAndWait();
 
         } catch (Exception e) {
-            LOGGER.error("Exception while load model view", e);
+            log.error("Exception while load model view", e);
         }
     }
