+++ src/main/java/com/tsel/home/project/booklibrary/controller/impl/book/EditBookController.java
@@ -167,7 +167,7 @@ public class EditBookController extends AbstractBookController {
 
     @FXML
     public void editBook() {
-        saveNewBook(bookForEdit.getAudiobookSites());
+        saveNewBook(bookForEdit.getId(), bookForEdit.getAudiobookSites());
     }
 
     @Override
