+++ src/main/java/com/tsel/home/project/booklibrary/controller/impl/book/NewBookController.java
@@ -128,7 +128,7 @@ public class NewBookController extends AbstractBookController {
 
     @FXML
     public void addBook() {
-        saveNewBook(audioBookSites);
+        saveNewBook(null, audioBookSites);
     }
 
     @FXML
